Crate diffuse_rs_common

Modules

core

nn

candle-nn

Macros

bail

map_dtype

test_device

Structs

NiceProgressBar

Nice progress bar with over an iterator and a message. COLOR is one of r,g,b

Enums

FileData

FileLoader

ModelSource

Source from which to load the model. This is easiest to create with the various constructor functions.

TokenSource

The source of the HF token.

Functions

conv1d

conv1d_no_bias

conv2d

conv2d_no_bias

embedding

from_mmaped_safetensors

Load tensors into a VarBuilder backed by a VarMap using MmapedSafetensors. Set silent to not show a progress bar.

get_token

This reads a token from a specified source. If the token cannot be read, a warning is logged with tracing and no token is used.

group_norm

layer_norm

linear

linear_b

linear_no_bias

load_bpe_tokenizer

Type Aliases

VarBuilder

A simple VarBuilder, this is less generic than VarBuilderArgs but should cover most common use cases.