diffuse_rs_common::core::error
Enum Error

pub enum Error {

    UnexpectedDType {
        msg: &'static str,
        expected: DType,
        got: DType,
    },
    DTypeMismatchBinaryOp {
        lhs: DType,
        rhs: DType,
        op: &'static str,
    },
    DTypeMismatchBinaryOp3 {
        lhs: DType,
        rhs: DType,
        c: DType,
        op: &'static str,
    },
    UnsupportedDTypeForOp(DType, &'static str),
    DimOutOfRange {
        shape: Shape,
        dim: i32,
        op: &'static str,
    },
    DuplicateDimIndex {
        shape: Shape,
        dims: Vec<usize>,
        op: &'static str,
    },
    UnexpectedNumberOfDims {
        expected: usize,
        got: usize,
        shape: Shape,
    },
    UnexpectedShape {
        msg: String,
        expected: Shape,
        got: Shape,
    },
    ShapeMismatch {
        buffer_size: usize,
        shape: Shape,
    },
    ShapeMismatchBinaryOp {
        lhs: Shape,
        rhs: Shape,
        op: &'static str,
    },
    ShapeMismatchCat {
        dim: usize,
        first_shape: Shape,
        n: usize,
        nth_shape: Shape,
    },
    ShapeMismatchSplit {
        shape: Shape,
        dim: usize,
        n_parts: usize,
    },
    OnlySingleDimension {
        op: &'static str,
        dims: Vec<usize>,
    },
    EmptyTensor {
        op: &'static str,
    },
    DeviceMismatchBinaryOp {
        lhs: DeviceLocation,
        rhs: DeviceLocation,
        op: &'static str,
    },
    DeviceMismatchBinaryOp3 {
        lhs: DeviceLocation,
        rhs: DeviceLocation,
        c: DeviceLocation,
        op: &'static str,
    },
    NarrowInvalidArgs {
        shape: Shape,
        dim: usize,
        start: usize,
        len: usize,
        msg: &'static str,
    },
    Conv1dInvalidArgs {
        inp_shape: Shape,
        k_shape: Shape,
        padding: usize,
        stride: usize,
        msg: &'static str,
    },
    InvalidIndex {
        op: &'static str,
        index: usize,
        size: usize,
    },
    BroadcastIncompatibleShapes {
        src_shape: Shape,
        dst_shape: Shape,
    },
    CannotSetVar {
        msg: &'static str,
    },
    MatMulUnexpectedStriding(Box<MatMulUnexpectedStriding>),
    RequiresContiguous {
        op: &'static str,
    },
    OpRequiresAtLeastOneTensor {
        op: &'static str,
    },
    OpRequiresAtLeastTwoTensors {
        op: &'static str,
    },
    BackwardNotSupported {
        op: &'static str,
    },
    NotCompiledWithCudaSupport,
    NotCompiledWithMetalSupport,
    CannotFindTensor {
        path: String,
    },
    Cuda(Box<dyn Error + Send + Sync>),
    Metal(MetalError),
    Ug(Error),
    TryFromIntError(TryFromIntError),
    Npy(String),
    Zip(ZipError),
    ParseInt(ParseIntError),
    FromUtf8(FromUtf8Error),
    Io(Error),
    SafeTensor(SafeTensorError),
    UnsupportedSafeTensorDtype(Dtype),
    Wrapped(Box<dyn Error + Send + Sync>),
    WrappedContext {
        wrapped: Box<dyn Error + Send + Sync>,
        context: String,
    },
    WithPath {
        inner: Box<Self>,
        path: PathBuf,
    },
    WithBacktrace {
        inner: Box<Self>,
        backtrace: Box<Backtrace>,
    },
    Msg(String),

}

Main library error type.

Variants

UnexpectedDType

Fields

msg: &'static str

expected: DType

got: DType

DTypeMismatchBinaryOp

Fields

lhs: DType

rhs: DType

op: &'static str

DTypeMismatchBinaryOp3

Fields

lhs: DType

rhs: DType

c: DType

op: &'static str

UnsupportedDTypeForOp(DType, &'static str)

DimOutOfRange

Fields

shape: Shape

dim: i32

op: &'static str

DuplicateDimIndex

Fields

shape: Shape

dims: Vec<usize>

op: &'static str

UnexpectedNumberOfDims

Fields

expected: usize

got: usize

shape: Shape

UnexpectedShape

Fields

msg: String

expected: Shape

got: Shape

ShapeMismatch

Fields

buffer_size: usize

shape: Shape

ShapeMismatchBinaryOp

Fields

lhs: Shape

rhs: Shape

op: &'static str

ShapeMismatchCat

Fields

dim: usize

first_shape: Shape

n: usize

nth_shape: Shape

ShapeMismatchSplit

Fields

shape: Shape

dim: usize

n_parts: usize

OnlySingleDimension

Fields

op: &'static str

dims: Vec<usize>

EmptyTensor

Fields

op: &'static str

DeviceMismatchBinaryOp

Fields

lhs: DeviceLocation

rhs: DeviceLocation

op: &'static str

DeviceMismatchBinaryOp3

Fields

lhs: DeviceLocation

rhs: DeviceLocation

c: DeviceLocation

op: &'static str

NarrowInvalidArgs

Fields

shape: Shape

dim: usize

start: usize

len: usize

msg: &'static str

Conv1dInvalidArgs

Fields

inp_shape: Shape

k_shape: Shape

padding: usize

stride: usize

msg: &'static str

InvalidIndex

Fields

op: &'static str

index: usize

size: usize

BroadcastIncompatibleShapes

Fields

src_shape: Shape

dst_shape: Shape

CannotSetVar

Fields

msg: &'static str

MatMulUnexpectedStriding(Box<MatMulUnexpectedStriding>)

RequiresContiguous

Fields

op: &'static str

OpRequiresAtLeastOneTensor

Fields

op: &'static str

OpRequiresAtLeastTwoTensors

Fields

op: &'static str

BackwardNotSupported

Fields

op: &'static str

NotCompiledWithCudaSupport

NotCompiledWithMetalSupport

CannotFindTensor

Fields

path: String

Cuda(Box<dyn Error + Send + Sync>)

Metal(MetalError)

Ug(Error)

TryFromIntError(TryFromIntError)

Npy(String)

Zip(ZipError)

Zip file format error.

ParseInt(ParseIntError)

Integer parse error.

FromUtf8(FromUtf8Error)

Utf8 parse error.

Io(Error)

I/O error.

SafeTensor(SafeTensorError)

SafeTensor error.

UnsupportedSafeTensorDtype(Dtype)

Wrapped(Box<dyn Error + Send + Sync>)

Arbitrary errors wrapping.

WrappedContext

Arbitrary errors wrapping with context.

Fields

wrapped: Box<dyn Error + Send + Sync>

context: String

WithPath

Adding path information to an error.

Fields

inner: Box<Self>

path: PathBuf

WithBacktrace

Fields

inner: Box<Self>

backtrace: Box<Backtrace>

Msg(String)

User generated error message, typically created via bail!.

Implementations

impl Error

pub fn wrap(err: impl Error + Send + Sync + 'static) -> Self

Create a new error by wrapping another.

pub fn msg<M: Display>(msg: M) -> Self

Create a new error based on a printable error message.

If the message implements std::error::Error, prefer using Error::wrap instead.

pub fn debug(err: impl Debug) -> Self

Create a new error based on a debuggable error message.

If the message implements std::error::Error, prefer using Error::wrap instead.

pub fn bt(self) -> Self

pub fn with_path<P: AsRef<Path>>(self, p: P) -> Self