diffuse_rs_common::core
Trait IndexOp

pub trait IndexOp<T> {
    // Required method
    fn i(&self, index: T) -> Result<Tensor, Error>;
}

Trait used to implement multiple signatures for ease of use of the slicing of a tensor

Required Methods

fn i(&self, index: T) -> Result<Tensor, Error>

Returns a slicing iterator which are the chunks of data necessary to reconstruct the desired tensor.