diffuse_rs_common::nn
Module batch_norm

Batch Normalization.

This layer applies Batch Normalization over a mini-batch of inputs as described in Batch Normalization. The input is expected to have at least three dimensions.

Note that this implementation is for inference only, there is no possibility to track the running stats.

Structs

BatchNorm

BatchNormConfig

Functions

batch_norm