diffuse_rs_common::nn::var_builder
Trait Renamer

pub trait Renamer {
    // Required method
    fn rename(&self, v: &str) -> Cow<'_, str>;
}

This traits specifies a way to rename the queried names into names that are stored in an inner VarBuilder.

Required Methods

fn rename(&self, v: &str) -> Cow<'_, str>

This is applied to the name obtained by a name call and the resulting name is passed to the inner VarBuilder.